+++ driver/src/perfcnt/xaie_perfcnt.c
@@ -772,7 +772,7 @@ AieRC XAie_PerfCounterGetControlConfig(XAie_DevInst *DevInst, XAie_LocType Loc,
 
 	/* Get start and stop event individually and store in event pointer */
 	RegEvent = StartStopEvent & PerfMod->Start.Mask;
-	RC = XAie_EventPhysicalToLogicalConv(DevInst, Loc, Module, (u8)RegEvent,
+	RC = XAie_EventPhysicalToLogicalConv(DevInst, Loc, Module, (u16)RegEvent,
 			StartEvent);
 	if (RC != XAIE_OK) {
 		return RC;
@@ -780,7 +780,7 @@ AieRC XAie_PerfCounterGetControlConfig(XAie_DevInst *DevInst, XAie_LocType Loc,
 
 	RegEvent = (StartStopEvent & PerfMod->Stop.Mask) >>
 			PerfMod->StartStopShift / 2U;
-	RC = XAie_EventPhysicalToLogicalConv(DevInst, Loc, Module, (u8)RegEvent,
+	RC = XAie_EventPhysicalToLogicalConv(DevInst, Loc, Module, (u16)RegEvent,
 			StopEvent);
 	if (RC != XAIE_OK) {
 		return RC;
@@ -798,7 +798,7 @@ AieRC XAie_PerfCounterGetControlConfig(XAie_DevInst *DevInst, XAie_LocType Loc,
 	/* Get reset event for given counter and store in the event pointer */
 	RegEvent = RegEvent >> Counter * PerfMod->ResetShift;
 	RegEvent &= PerfMod->Reset.Mask;
-	RC = XAie_EventPhysicalToLogicalConv(DevInst, Loc, Module, (u8)RegEvent,
+	RC = XAie_EventPhysicalToLogicalConv(DevInst, Loc, Module, (u16)RegEvent,
 			ResetEvent);
 	if (RC != XAIE_OK) {
 		return RC;
